Move reindexContacts to shared file and add OpenSearch reset to tests
