+++ core/tasks/ctasks/urn_added_test.go
@@ -18,7 +18,7 @@ func TestURNAdded(t *testing.T) {
 	vc := rt.VK.Get()
 	defer vc.Close()
 
-	defer testsuite.Reset(t, rt, testsuite.ResetData|testsuite.ResetValkey)
+	defer testsuite.Reset(t, rt, testsuite.ResetData|testsuite.ResetValkey|testsuite.ResetOpenSearch)
 
 	// add a new URN that doesn't exist in the database
 	err := tasks.QueueContact(ctx, rt, testdb.Org1.ID, testdb.Ann.ID, &ctasks.URNAdded{URN: "tel:+16055749999"})
